change feature data type from int16 to int32

+++ src/fastlbp_imbg/common.py
@@ -1,3 +1,3 @@
 import numpy as np
 
-_features_dtype = np.uint16
+_features_dtype = np.uint32
